ML-service/requirements.txt
nemo-toolkit==1.21.0